Quit sharing your work
You guys can help each other, but do not copy directly. Make sure you try to solve the problem and understand what the program is doing. If you are sharing work, make sure those you share it with have tried to solve it first and that after helping them that they understand what the program is doing.